+++ utilities/inc/gutenberg/src/index.js
@@ -1,11 +1,14 @@
 import { dispatch } from '@wordpress/data';
 import apiFetch from '@wordpress/api-fetch';
-import { addFilter } from '@wordpress/hooks';
-import classnames from "classnames";
+import { registerFormatType, toggleFormat } from '@wordpress/rich-text';
+import { RichTextToolbarButton } from '@wordpress/block-editor';
+import { __, _n } from '@wordpress/i18n';
+import { lineDashed } from '@wordpress/icons';
 
 // remove template panel
-dispatch( 'core/editor').removeEditorPanel( 'template' );
+// dispatch( 'core/editor').removeEditorPanel( 'template' );
 
+// set background color
 apiFetch( { path: 'gutenberg/v1/getBackgroundColor' } ).then( ( color ) => {
     const $sheet = document.createElement('style' );
 
@@ -18,9 +21,7 @@ apiFetch( { path: 'gutenberg/v1/getBackgroundColor' } ).then( ( color ) => {
     document.body.appendChild( $sheet );
 } );
 
-/**
- * Set client width property to get fluid to work properly.
- */
+// set client width property to get fluid to work properly
 (function setClientWidthProperty() {
     const $editorStylesWrapper = document.querySelector( '.editor-styles-wrapper' );
     if ( $editorStylesWrapper ) new ResizeObserver( entries => {
@@ -28,3 +29,32 @@ apiFetch( { path: 'gutenberg/v1/getBackgroundColor' } ).then( ( color ) => {
     } ).observe( $editorStylesWrapper );
     else window.requestAnimationFrame( setClientWidthProperty )
 })();
+
+// `style="white-space: nowrap"`
+registerFormatType( 'utilities-gutenberg/nowrap', {
+    title: __( 'no line break', 'gutenberg' ),
+    tagName: 'span',
+    className: null,
+    attributes: {
+        style: 'style'
+    },
+    edit: ( { isActive, onChange, value } ) => {
+        return (
+            <RichTextToolbarButton
+                icon={ lineDashed }
+                title={ __( 'no line break', 'gutenberg' ) }
+                onClick={ () => {
+                    onChange(
+                        toggleFormat( value, {
+                            type: 'utilities-gutenberg/nowrap',
+                            attributes: {
+                                style: 'white-space: nowrap'
+                            }
+                        } )
+                    );
+                } }
+                isActive={ isActive }
+            />
+        );
+    },
+} );