+++ utilities/inc/plugins/contact-form-7/inc/mail-sent-ok-redirect/admin.js
@@ -1,0 +1,7 @@
+(function( window, document, $, undefined ) {
+
+    $( '#wpcf7-message-mail-sent-ok' )
+        .suggest( ajaxurl + '?action=suggest-link-ajax&page=1' )
+        .after( '<br />', `<span>${wp.i18n.__( 'If you enter a URL, the user will be redirected to it.', 'cf7' )}</span>` );
+
+})( window, document, jQuery );