+++ utilities/inc/security/inc/updates/updates.php
@@ -1,5 +1,12 @@
 <?php
 
+// disable auto updates
+add_filter( 'allow_minor_auto_core_updates', '__return_false' );
+add_filter( 'allow_major_auto_core_updates', '__return_false' );
+add_filter( 'auto_update_plugin', '__return_false' );
+add_filter( 'auto_update_theme', '__return_false' );
+add_filter( 'auto_update_translation', '__return_false' );
+
 /**
  * Dashboard's security widget entry.
  */